requirements.txt
streamlit
pandas
numpy
yfinance
plotly
scipy
altair
pillow